Add minifying back to production build and avoid mangling function names
We use constructor.name === 'SomeThing' tests in our code to identyfy which class a certain object comes from. Such tests will fail when code is minified because function names are lost when mangling. This tells uglify plugin to avoid mangling function names.

React uses the presence of process.env.NODE_ENV set to 'production' to remove warnings and stuff in production builds. So add that to define plugin.