Change connection retries number
This commit changes the retries number to unlimited. So the connection to the ws will be tried to establish until the debugging is stopped by the user. These retries will be tried using the executeOnPooledThread method. So IDEA will not hang. Additionally, this commit changes the sendBreakpoints method as well.