Prevent format in source view upon syntax error
As we are using the parsed syntax tree to format
source, when there are syntax errors - if somehow
parser recovered from the error (as we allow so
with our error strategy), there is a possibility
that parser sends us an incomplete AST to which
the formatted source will be generated. That means
when user do a format in a such situation, he
will get a newly formatted source with some pieces
missing from it.