Seperate source-editor from source-view
ATM, source view is only rendered once &
will avoid react re-render (with the help
of shouldComponentUpdate). This is because
ace editor doesn't support react. However,
with the decision to show no of syntax errors
& related behaviours on top of design-view switch
btn (and ease of using react to implement it), we
now want source-view to be able to re-rendered from
react. Hence, seperating ace-editor related stuff
to a seperate component down the hierarchy.