Make plugins capable of emitting events
When a plugin needs to communicate with one of its views
(eg. notify a view when something happens so that it can update
, etc.), it's not good to use app level commands always as it is
not something meant to be seen by others (it is internal). To
support this, we need to make plugins capable of emitting events
which it's view can subscribe to. Since plugin ref is not visible
to other plugins (only the public context of a plugin is visible
to others), this won't leed to abuse of plugin events to do hacks.