Fix undo redo / dirty state glitches
The issue was happening because setContent call in the undo manager triggers a rerender in source editor; and source editor replaces the content again.
But this replacing does not trigger the execute method of undo manager, possibly because the content is the same. So the state we set to skip undo manager is never unset.