Prevent opening of program dir upon sample open
Prevent it for samples which have a package declared.
Composer will automatically popup a dialog asking
user's confirmation to open the program dir in such
cases. Hence we need to stop duplicated task of opening the
folder here. However, there are some samples which doesn't
have a package declared but contains multiple files and
we need to open the relavant folder
while opening the sample. Hence, I did not remove all the logic
for opening folder while opening samples. Instead removed that
feature from the samples which we know have packages and let
the automatic wizard handles that as it does for other files.