Make sure back end is up before attempting js tests.
Previously the back end composer services was started as a async maven
execution. The services may or may not be up when starting js tests.

By starting them inside the test script itself and waiting for services
stdout to print the success message, its guaranteed that the back end is
up when starting js tests.