Make all functions private
There is no need for the micro-transaction functions or structs to
be public because it is only called by internal Ballerina code and
never called from outside. The functions were made public temporarily
for testing purposes and now I am making them private after implementing
the rest of the stuff.