Improve performance of the LS by reusing the package cache
By design PackageCache is not exposed to the outside. Thus, exposed it
using setInstance() method. Whenever LS receives a completion request a
code compilation is triggered with a new Compiler Context. The
PackageCache is bound to the Compiler Context. Thus, per each request
compiler has to load external packages.

This fix improves the performance of the LS by reusing the PackageCache.
The approach used here is to inject a global PackageCache into the
CompilerContext. Added some extension points to the PackageCache to
inject any external caching approach.