Write codegen output into ballerina package
Ballerina now supports projects and packages to organize the
code. Since most of the codegen use cases will involve in
manipulation of generated code we will be using projects and
packages from now on, to generate the code.
Currently 'ballerina swagger' command should run within a
project dir or it'll fail.