Prevent non-existent error popup during file rename
When an already opened file in editor is renamed from
file explorer, it will move the file in fs and open
a new editor tab for the file. Then it will close existing
editor. However, a new validation added to composer for
verifing file's existency in fs has caused a regression
issue with this approach. This PR fixes the regression.

Issue: https://github.com/ballerina-platform/ballerina-lang/issues/8550