Fix breakpoints are not being deployed correctly
Breakpoints sent from the Debug Client are not correctly being deployed
for the projects structures as mentioned in https://github.com/ballerina-platform/ballerina-lang/issues/10056. This is due to the
fact that `packagesInfoMap` inside the DebugInfoHolder contains the built
packages without the `organization` and `version` whenever a Ballerina.toml
presents.

Therefore, added a fix to add the version and the organization into the top level
built package. Versions of the package imports inside the code will also
the project version through the enclosure package.

This commit resolves https://github.com/ballerina-platform/ballerina-lang/issues/10056