Avoid redundant bundling of LS classes in swagger utils jar
ATM LS classes are being duplicated due to repackaging them
within uber jar for swagger-to-ballerina utils. This is
because of swagger util's dependency to LSCompiler. To solve,
we decided to move LS compiler jar into bre/lib where it will
be available in classpath when ballerina swagger command is executed.
Hence the need for repackaging goes away and since we anyway
want bre/lib in class path to start LS, we don't need any
change to that path as well.