Move monaco based lang-client to a separate module.
Currently, the lang-client implementation based on
monaco-langclient - which was implemented with the
need for having a lang-server client in test envs
without having to configure it as a vscode ext -
is inside integration-tests module. But this needs
to be reused in other modules such as diagram for
storybooking needs while development. Hence the
seperation.