Mount endpoint search dialog on diagram container rather on body
In vs-code, since diagram is rendered on top of an Iframe, it
doesn't matter that semantic ui's modals are mounted on the
default node which is body. But in other cases such as Theia,
the body is the root of the whole editor and it makes modals
appear in random places. This will make sure it is mounted
on diagram container everywhere.