Refactor bvm with adding size() default method
We have used length and size functions in Bvalues. When getting the length we need to filter either length or size method to calculate length. With default method with size() this will be generalize, so we do not need to check the type. This will increase the performance as well.