Remove the need to have the obj directory and support other orgs
The obj the directory which has the balos was needed to run any module. This change removes that requirement.
Also the gradle files have been changed to use the org name from the Ballerina.toml file to properly copy the files if the org name is different from ballerina. In this case the toml file (in /src/main/ballerina/Ballerina.toml folder) should have the org name.
Moreover there's is an issue with the project hierarchy not being properly implemented which causes ballerinax to be not recognized. It is now hardcoded in the BootstrapRunner class. This has to be removed and fixed.