Improve filler value check for Union with BFinite
we create a union with constant values it creates a BFinite type. When
comparing BFinite with current BSameTypeVisitor compares BFinite types
with value instead of instead of type.

For filler value check BFinite will be checked to have same types like
int. No changes done to BSameTypeVisitor and the logic is separately
handled

Signed-off-by: KRVPerera <rukshan.viduranga@gmail.com>