Issue error on push/pop call on fixed-length lists
Push/Pop functions do not natively support annotation or a way to
declare that they will change the length of the list. This is not
allowed in fixed-length arrays.

`push` and `pop` names are hardcoded and checked against invocations on
fixed-length arrays and tuples without rest parameters to identify the
invalid calls on compile time.

Signed-off-by: KRVPerera <rukshan.viduranga@gmail.com>