Fix taintedness propagation of object functions
Previously if the receiver is tainted any member function
would return tainted value regardless of receivers involment
in the return value. Now if the receiver is tainted and
the taint table says that when receiver is tainted the return
value is tainted, then we mark the return value of a member
tainted.