Fix issue when calling function with error return in service
The issue is in generating the method call of a function in a service. Earlier since only resources were allowed in a service the method calls of all functions in a service were throwing errors when it has an error return. Now a check has been added to throw an error only for a resource.
Resolve https://github.com/ballerina-platform/ballerina-lang/issues/24795