Add map having the local variables used in lamdas to table
At ballerina desugar level a map is introduced to hold the variables passed to a lambda. These variables as well as the map holding them are not in the LocalVariableTable. This prevents the debugger from showing them. This PR adds the map to the LocalVariableTable so that the variables in it are accessible to the debugger.

Related issues: https://github.com/ballerina-platform/ballerina-lang/issues/27739 and https://github.com/ballerina-platform/ballerina-lang/issues/27738