Remove return statement from LVT
The return statement is in a separate Basic Block(BB). When the program has multiple return statements they all go to the same BB. In this case the position of the actual return statement in the bal code is unclear. The actual return statements are indicated by the GOTO statements to the BB with the return statement. So only the GOTO statements are included in the LVT and the return statements are ignored.

Resolve https://github.com/ballerina-platform/ballerina-lang/issues/31061