Change new vars and dclns getter
Introduced two new attributes to ClassLoadInvoker namely newDefinedVariableNames and newModuleDeclnNames to hold information on the current executed code snippet. These values need to be removed before any new invocation because each snippet will use this to hold info on defined vars for it. These are for the use of BalShellService to collect info on executed code snippet.