Add boolean flag to track transaction abort status in endXATransaction()
This commit introduces a boolean variable to keep track of whether the transaction is set for abort only or not. The variable indicate the success or failure of the work carried out within the transaction context. This change aims to provide a better control over the transactional workflow, enhancing the handling of abort scenarios within the XA transaction management.

If the transaction is aborting the xa_end() is called on the XA Resource with the TMFAIL flag indicating that the  portion of work has failed marking the transaction as rollback-only. Otherwise, the TMSUCCESS flag will be used as per normal operation.