Changes to the recovery logic
This commit changes the initial logic where you check the log and take actions according to that on xa resource to other way around by getting the failed transactions in xa resources and taking the action from the log. This way we can ensure that there are no unnecessary commit/rollback calls made to the xa resources which would return invalid xid errors.