Update jline version for test compatibility
jline 3.25.1 which is the latest at time of writing, breaks most of our
test cases related ballerina shell. After testing around it seems that
this is caused by the dumbterminal. Even though tests failed the bal
shell command seems to work as usual. There is no mention of this issue
on jline's issues and I was not able to reproduce it in a sample java
project. The version bump I needed to access jansi from within jline
happened in jline 3.25.0, therefore instead of 3.25.1 I changed it back
to 3.25.0 until we find a way to change our test cases or jline's
future updates.