Fix expected error message
Fix runtime type clash

Due to reasons that are not entierly clear to me when we run all the
unit tests some how types from `record_project_closed_rec_equiv` to
clashed with types defined in other unit tests. As a temperary
workaround I introduced a new package name for this.