Fix wrong interop error msg for error included return types
Previously ballerina `readonly` return type allowed for java void methods. However, `error?` was not allowed with java void. This inconsistency is due to type checking not recognizing the fact that readonly type does include the error type. The error given in the latter case too is wrong because error is optional.